Fixing a regression from changeset 3074 (c867117688bb): Reordering contacts and groups was broken.  Now it's fixed.

+++ Frameworks/Adium Framework/Source/AIListObject.h
@@ -91,8 +91,8 @@ typedef enum {
 	//Grouping, Manual ordering
 	NSMutableSet *m_groups; //The AIContainingObjects that this object is in
 	
-	CGFloat				cachedSmallestOrder;
-	CGFloat				cachedLargestOrder;
+	float				cachedSmallestOrder;
+	float				cachedLargestOrder;
 }
 
 - (id)initWithUID:(NSString *)inUID service:(AIService *)inService;
@@ -174,10 +174,10 @@ typedef enum {
  * but are implemented by AIListObject (which does not conform to the protocol) for the convenience
  * of subclasses.
  */
-@property (readonly, nonatomic) CGFloat smallestOrder;
-@property (readonly, nonatomic) CGFloat largestOrder;
-- (void)listObject:(AIListObject *)listObject didSetOrderIndex:(CGFloat)inOrderIndex;
-- (CGFloat)orderIndexForObject:(AIListObject *)listObject;
+@property (readonly, nonatomic) float smallestOrder;
+@property (readonly, nonatomic) float largestOrder;
+- (void)listObject:(AIListObject *)listObject didSetOrderIndex:(float)inOrderIndex;
+- (float)orderIndexForObject:(AIListObject *)listObject;
 - (void) moveContainedObject:(AIListObject *)object toIndex:(NSInteger)index;
 
 @end
